+++ sdk/evaluation/azure-ai-evaluation/azure/ai/evaluation/_evaluators/_content_safety/_sexual.py
@@ -1,10 +1,21 @@
 # ---------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
+<<<<<<< HEAD
 from typing_extensions import override
 from azure.ai.evaluation._common.constants import EvaluationMetrics
 from azure.ai.evaluation._model_configurations import AzureAIProject
 from azure.ai.evaluation._evaluators._common._base_rai_svc_eval import BaseRaiServiceEval
+=======
+from promptflow._utils.async_utils import async_run_allowing_running_loop
+
+from azure.ai.evaluation._common.constants import EvaluationMetrics
+
+try:
+    from ._content_safety_base import ContentSafetyEvaluatorBase
+except ImportError:
+    from _content_safety_base import ContentSafetyEvaluatorBase
+>>>>>>> main
 
 
 class SexualEvaluator(BaseRaiServiceEval):